+++ src/geometry/circle.js
@@ -47,9 +47,7 @@ TYPE6.Geometry.Circle = {
     var obj = Object.create(this);
     obj.init();
     obj.setRadius(radius);
-    obj.initSize();
     obj.setPositionXY(positionX, positionY);
-    
     return obj;
   },
 
@@ -62,11 +60,17 @@ TYPE6.Geometry.Circle = {
     this.position = TYPE6.Vector2D.create();
     this.radius   = 0.0;
     this.diameter = 0.0;
+    this.initSize();
   },
   
   initSize: function(){
-    this.size     = TYPE6.Vector2D.create( this.diameter, this.diameter );
-    this.halfSize = TYPE6.Vector2D.create( this.radius, this.radius );
+    this.size     = TYPE6.Vector2D.create();
+    this.halfSize = TYPE6.Vector2D.create();
+  },
+  
+  setSize: function(){
+    this.size.setXY( this.diameter, this.diameter );
+    this.halfSize.setXY( this.radius, this.radius );
   },
 
   /**
@@ -185,7 +189,7 @@ TYPE6.Geometry.Circle = {
   setRadius: function( radius ){
     this.radius   = radius;
     this.diameter = this.radius * 2;
-    this.initSize();
+    this.setSize();
     return this.radius;
   },
 
@@ -210,7 +214,7 @@ TYPE6.Geometry.Circle = {
   setDiameter: function( diameter ){
     this.diameter = diameter;
     this.radius   = this.diameter * 0.5;
-    this.initSize();
+    this.setSize();
     return this.diameter;
   },
 
@@ -226,14 +230,23 @@ TYPE6.Geometry.Circle = {
   },
   
   /**
-  * get size.
-  * @since 0.2.0
+  * get halfSize.
+  * @since 0.2.1
   * @method
-  * @param {array(2)} An array of floats with array[0] as x and array[1] as y.
   * @returns {Vector2D}
   */
   getHalfSize: function(){
     return this.halfSize;
+  },
+  
+  /**
+  * get size.
+  * @since 0.2.1
+  * @method
+  * @returns {Vector2D}
+  */
+  getSize: function(){
+    return this.size;
   },
 
   /**
@@ -299,11 +312,11 @@ TYPE6.Geometry.Circle = {
   * @since 0.2.3
   * @method
   * @param {context} context The context of the canvas.
-  * @param {string} color The color of the circle.
-  * @returns {Vector2D}
+  * @param {string} fillColor The fill color of the circle.
+  * @param {string} strokeColor The stroke color of the circle.
+  * @param {float} strokeWidth The stroke width of the circle.
   */
-  draw: function( context, color ){
-    context.fillStyle = color;
+  draw: function( context, fillColor, strokeColor, strokeWidth ){
     context.beginPath();
     context.arc(  this.getPositionX(),
                   this.getPositionY(),
@@ -312,7 +325,15 @@ TYPE6.Geometry.Circle = {
                   TYPE6.Trigonometry.TWOPI,
                   false
                 );
-    context.fill();
+    if( fillColor ){
+      context.fillStyle = fillColor;
+      context.fill();
+    }
+    if( strokeColor ){
+      context.strokeStyle = strokeColor;
+      context.lineWidth = strokeWidth;
+      context.stroke();
+    }
   }
 
 };