+++ website/js/examples/geometry/clamp.js
@@ -20,14 +20,14 @@ window.onload = function() {
     context.clearRect(0, 0, width, height);
 
     drawRectangles();
-    circle.draw( context, "#CC0000" );
+    circle.draw( context, "#CC0000", null, null );
 
-    context.fill();
+    //context.fill();
   });
 
   function drawRectangles(){
-    rect.draw( context, "#cccccc" );
-    innerRect.draw( context, "#999999" );
+    rect.draw( context, "#cccccc", null, null );
+    innerRect.draw( context, "#999999", "#000000", 1 );
   }
 
 };