+++ website/public/js/examples/geometry/distance.js
@@ -1,0 +1,1 @@
+../../../../js/examples/geometry/distance.js