+++ website/public/js/examples/mathUtils/lerp.js
@@ -1,0 +1,1 @@
+../../../../js/examples/mathUtils/lerp.js