+++ website/public/js/examples/mathUtils/map.js
@@ -1,0 +1,1 @@
+../../../../js/examples/mathUtils/map.js