+++ website/public/js/examples/random/2d-random-distribution.js
@@ -1,0 +1,1 @@
+../../../../js/examples/random/2d-random-distribution.js